// resources/ts/app/package/modules/uuid/types.ts
/*****************************************
 * Package Module UUID Generator
 *
 * Types
 *****************************************/

/*----------------------------------------*
 * Version 7
 *----------------------------------------*/

/**
 * V7 State
 */
export type V7State = {
    /**
     * milliseconds state
     *
     * @type {number}
     */
    milliseconds: number;

    /**
     * sequence state
     *
     * @type {number}
     */
    sequence: number;
};